requirements.txt
torch
torchvision
numpy
Pillow
nibabel